Migrate to new architecture
