+++ Examples/IconExplorer/ios/.xcode.env
@@ -3,6 +3,15 @@
 # To customize your local environment, you can create an `.xcode.env.local`
 # file that is not versioned.
 
+# Set up asdf-vm if present
+ if [[ -f "$HOME/.asdf/asdf.sh" ]]; then
+   # shellcheck source=/dev/null
+   . "$HOME/.asdf/asdf.sh"
+ elif [[ -x "$(command -v brew)" && -f "$(brew --prefix asdf)/asdf.sh" ]]; then
+   # shellcheck source=/dev/null
+   . "$(brew --prefix asdf)/asdf.sh"
+ fi
+
 # NODE_BINARY variable contains the PATH to the node executable.
 #
 # Customize the NODE_BINARY variable here.
