--- Examples/IconExplorer/scripts/postinstall.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-rm -rf node_modules/react-native-vector-icons/Examples
-rm -rf node_modules/react-native-vector-icons/node_modules
