+++ RNVectorIconsManager/RNVectorIconsManager.h
@@ -6,16 +6,8 @@
 //  Copyright (c) 2015 Joel Arvidsson. All rights reserved.
 //
 
-#if __has_include(<React/RCTBridgeModule.h>)
 #import <React/RCTBridgeModule.h>
-#else // Compatibility for RN version < 0.40
-#import "RCTBridgeModule.h"
-#endif
-#if __has_include(<React/RCTLog.h>)
 #import <React/RCTLog.h>
-#else // Compatibility for RN version < 0.40
-#import "RCTLog.h"
-#endif
 
 FOUNDATION_EXPORT NSString *const RNVIErrorDomain;
