+++ lib/create-icon-set.js
@@ -1,19 +1,11 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import {
-  NativeModules,
-  Platform,
-  PixelRatio,
-  processColor,
-  Text,
-} from 'react-native';
+import { Platform, PixelRatio, processColor, Text } from 'react-native';
 
 import ensureNativeModuleAvailable from './ensure-native-module-available';
 import createIconSourceCache from './create-icon-source-cache';
 import createIconButtonComponent from './icon-button';
-
-export const NativeIconAPI =
-  NativeModules.RNVectorIconsManager || NativeModules.RNVectorIconsModule;
+import NativeIconAPI from './NativeRNVectorIcons';
 
 export const DEFAULT_ICON_SIZE = 12;
 export const DEFAULT_ICON_COLOR = 'black';
